[hotfix] Introduce ExecutionGraphHandler to factor out ExecutionGraph logic from the SchedulerBase
